Further Functionality
More updates to all functionality. Currently just missing adding in modal works as per our design.
Most of the code is based on the example project that was given to us in the LAMP Stack video, with adjustments made for our purposes. Several things still need to be adjusted as the html files are developing.